Refactor small/large nav bar into single class
